requirements.txt
Django==2.0.5
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.3
pytz==2018.4
six==1.11.0
